dealapp-scraper/requirements.txt
playwright==1.40.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.1.4